Summary: Fixed Find.cpp for 64 bit build Reviewer: bravindra, done Additional reviewers: Details: Fixed Find.cpp for 64 bit build
git-svn-id: svn://svn.code.sf.net/p/framewave/code/trunk@213 f364fc6c-7e41-0410-b16c-eb8c68fe7df8

+++ Framewave/domain/fwSignal/src/Find.cpp
@@ -32,7 +32,7 @@ namespace OPT_LEVEL
 
                 if(strend == str)
                 {
-                    index = ((int)src - lenfind) - (int)start ;
+                    index = (int)((src - lenfind) - start) ;
                     return index;
                 }
 
@@ -61,7 +61,7 @@ namespace OPT_LEVEL
 
                 if(strend == str && *strend == *src)
                 {
-                    index = ((int)(end- start) - lenfind + 1) ;
+                    index = (int)((end- start) - lenfind + 1) ;
                     return index;
                 }
             }
